fix: ui error (#144)
* fix: layout

* fix: layout

* fix: layout

* fix: import

* fix: mf-5705

* fix: mf-5701

* chore: prettier

* fix: lock file

* fix: revert lock file
