+++ src/app/(normal)/detail/[platform]/[id]/page.tsx
