+++ src/app/(normal)/detail/layout.tsx
