+++ src/app/(normal)/notifications/page.tsx
