+++ src/app/(normal)/profile/components/FarcasterProfile.tsx
@@ -2,24 +2,24 @@ import { useQuery } from '@tanstack/react-query';
 import { useMemo } from 'react';
 import { useDocumentTitle } from 'usehooks-ts';
 
-import ContentTabs from '@/app/profile/components/ContentTabs.js';
-import EmptyProfile from '@/app/profile/components/EmptyProfile.js';
-import Info from '@/app/profile/components/Info.js';
-import Title from '@/app/profile/components/Title.js';
+import ContentTabs from '@/app/(normal)/profile/components/ContentTabs.js';
+import EmptyProfile from '@/app/(normal)/profile/components/EmptyProfile.js';
+import Info from '@/app/(normal)/profile/components/Info.js';
+import Title from '@/app/(normal)/profile/components/Title.js';
 import Loading from '@/components/Loading.js';
 import { createPageTitle } from '@/helpers/createPageTitle.js';
 import { useLogin } from '@/hooks/useLogin.js';
 import { usePlatformAccount } from '@/hooks/usePlatformAccount.js';
-import { LensSocialMedia } from '@/providers/lens/SocialMedia.js';
+import { WarpcastSocialMedia } from '@/providers/warpcast/SocialMedia.js';
 
-interface LensProfileProps {
-    handle: string;
+interface FarcasterProfileProps {
+    id: string;
 }
-export default function LensProfile({ handle }: LensProfileProps) {
-    const lensClient = new LensSocialMedia();
+export default function FarcasterProfile({ id }: FarcasterProfileProps) {
+    const farcasterClient = new WarpcastSocialMedia();
     const { data: profile, isLoading } = useQuery({
-        queryKey: ['profile', handle],
-        queryFn: () => lensClient.getProfileByHandle(`lens/${handle}`),
+        queryKey: ['profile', id],
+        queryFn: () => farcasterClient.getProfileById(id),
     });
 
     const isLogin = useLogin();
@@ -27,14 +27,14 @@ export default function LensProfile({ handle }: LensProfileProps) {
     const platformAccount = usePlatformAccount();
 
     const isMyProfile = useMemo(
-        () => !!isLogin && platformAccount.lens?.handle === handle,
-        [handle, isLogin, platformAccount.lens?.handle],
+        () => !!isLogin && platformAccount.farcaster.id === id,
+        [id, isLogin, platformAccount.farcaster.id],
     );
 
     const title = useMemo(() => {
         if (!profile) return '';
         const fragments = [profile.displayName];
-        if (profile.handle) fragments.push(`(@${profile.handle})`);
+        if (profile.profileId) fragments.push(`(@${profile.profileId})`);
         return createPageTitle(fragments.join(' '));
     }, [profile]);
